+++ src/types/any.js
@@ -1,1 +1,3 @@
-export default class Any {};
+import Primitive from './primitive';
+
+export default class Any extends Primitive {};