+++ src/types/primitive.js
@@ -1,0 +1,7 @@
+import { valueOf } from '../meta';
+
+export default class Primitive {
+  get state() {
+    return valueOf(this);
+  }
+};