+++ tests/recursive.test.js
@@ -1,11 +1,27 @@
-describe.skip('recursive microstates', () => {
-  it('can handle recursive lists', function() {
+import expect from 'expect';
 
-  });
-  it('can handle recursive objects', function() {
+import { Any } from '../src/types';
+import { create } from '../src/microstates';
+import { valueOf } from '../src/meta';
 
+describe('recursive microstates', () => {
+
+  class Cons {
+    car = Any;
+    cdr = Cons;
+  }
+
+  let cons;
+  beforeEach(function() {
+    cons = create(Cons, { car: 5});
+  })
+
+  it('doesnt blow up', function() {
+    expect(cons).toBeDefined();
+    expect(valueOf(cons.car)).toEqual(5);
   });
+
   it('can handle recursive things', function() {
-
+    expect(cons.cdr.cdr.cdr).toBeDefined();
   });
 })